deck users lightbox w typahead, better design

+++ app/javascript/card-maker/components/deck-users-lightbox.jsx
@@ -57,13 +57,19 @@ class DeckUsersLightbox extends React.Component {
     }
   }
 
-  removeUser = (id) => {
+  removeUser = (id, name) => {
     if (id && id > 0) {
-      $.ajax({
-        url: cardMakerUrl(`decks/${this.props.deck.id}/users/${id}`),
-        type: 'DELETE',
-        success: this.reload
-      });
+      let ok = confirm(I18n.t('react.card_maker.are_you_sure_remove_user', {
+        userName: name 
+      }));
+
+      if (ok) {
+        $.ajax({
+          url: cardMakerUrl(`decks/${this.props.deck.id}/users/${id}`),
+          type: 'DELETE',
+          success: this.reload
+        });
+      }
     }
   }
 
@@ -76,57 +82,76 @@ class DeckUsersLightbox extends React.Component {
         contentLabel={I18n.t('react.card_maker.manage_deck_users')}
         parentSelector={() => {return document.getElementById('Page')}}
         bodyOpenClassName='noscroll'
-        className={`${styles.lightbox} ${styles.lightboxDeckUsers}`}
+        className={[styles.lNewLightbox, styles.lNewLightboxDeckUsers].join(' ')}
         overlayClassName={`fixed-center-wrap disable-overlay`}
       >
-        <h2>Manage users for deck <strong>{this.props.deck.name}</strong></h2>
-        <h2>Owner: {this.state.owner.userName}</h2>
-        <div className={styles.lightboxDeckUsersList}>
-          <div>Users:</div>
-          <ul>
-            { this.state.users.map((u) => {
-              return (
-                <li key={u.id}>
-                  {u.userName}&nbsp;
-                  <i onClick={ () => this.removeUser(u.id) } 
-                    className={`fa fa-lg fa-minus-circle ${styles.deckUserMinus}`}
-                  />
-                </li>
-              );
-            })}
-          </ul>
+        <div className={styles.deckUsersHead}>
+          <h2>Manage users for deck <strong>{this.props.deck.name}</strong></h2>
+          <h2>{I18n.t('react.card_maker.owner_user', {
+            userName: this.state.owner.userName 
+          })}</h2>
         </div>
-        <ReactAutocomplete
-          items={this.state.typeaheadOptions}
-          getItemValue={item => item.label}
-          onChange={e => {
-            let query = e.target.value;
-            this.setState({
-              typeaheadValue: query
-            });
-
-            if (query.length) {
-              $.getJSON('/users/typeahead/' + query, (data) => {
-                this.setState({
-                  typeaheadOptions: data
-                });
+        <div className={styles.lNewCol}>
+          <div className={styles.lightboxDeckUsersList}>
+            <h3 className={styles.deckUsersTitle}>{I18n.t('react.card_maker.users')}</h3>
+            <ul className={styles.deckUsers}>
+              { this.state.users.map((u) => {
+                return (
+                  <li key={u.id} className={styles.deckUser}>
+                    <span className={styles.deckUserName}>{u.userName}</span>
+                    <i onClick={ () => this.removeUser(u.id, u.userName) } 
+                      className={`fa fa-lg fa-minus-circle ${styles.deckUserMinus}`}
+                    />
+                  </li>
+                );
+              })}
+            </ul>
+          </div>
+        </div>
+        <div className={styles.lNewCol}>
+          <ReactAutocomplete
+            items={this.state.typeaheadOptions}
+            getItemValue={item => item.label}
+            onChange={e => {
+              let query = e.target.value;
+              this.setState({
+                typeaheadValue: query
               });
+
+              if (query.length) {
+                $.getJSON('/users/typeahead/' + query, (data) => {
+                  this.setState({
+                    typeaheadOptions: data
+                  });
+                });
+              }
+            }}
+            onSelect={(val, item) => {
+              this.setState({
+                typeaheadValue: val,
+                selectedUserId: item.id
+              });
+            }}
+            renderItem={(item, isHighlighted) =>
+              <div style={{ background: isHighlighted ? 'lightgray' : 'white' }}>
+                {item.label}
+              </div>
             }
-          }}
-          onSelect={(val, item) => {
-            this.setState({
-              typeaheadValue: val,
-              selectedUserId: item.id
-            });
-          }}
-          renderItem={(item, isHighlighted) =>
-            <div style={{ background: isHighlighted ? 'lightgray' : 'white' }}>
-              {item.label}
-            </div>
-          }
-          value={this.state.typeaheadValue}
-        />
-        <input onClick={this.addSelectedUser} type="submit" value="add user"/>
+            renderInput={(props) =>
+              <input type='text' className={[styles.newInput, styles.newInputTxt].join(' ')} {...props} />
+            }
+            renderMenu={function(items, value, style)  {
+              return <div style={{ ...style, ...this.menuStyle, zIndex: 10 }} children={items}/>
+            }}
+            value={this.state.typeaheadValue}
+            wrapperStyle={{}}
+          />
+          <div>
+            <button className={styles.createBtn} onClick={this.addSelectedUser}>
+              {I18n.t('react.card_maker.add_user')}
+            </button>
+          </div>
+        </div>
       </ReactModal>
     );
   }
